requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-pinecone==0.0.3
pinecone-client==3.0.0
PyPDF2==3.0.1
streamlit==1.29.0
python-dotenv==1.0.0
tiktoken==0.5.2